module3-task1: * merge mock helpers
