+++ src/mocks/helpers/generate-artist-question-answers.helper.ts
@@ -1,5 +1,13 @@
-import { generateArtistQuestionAnswer } from '~/mocks/helpers/generate-artist-question-answer.helper';
 import { IArtistQuestionAnswer } from '~/common/interfaces/interfaces';
+
+const AVATAR_URL = `https://api.adorable.io/avatars/128`;
+
+const generateArtistQuestionAnswer = (
+  artist: string
+): IArtistQuestionAnswer => ({
+  artist,
+  picture: `${AVATAR_URL}/${Math.random()}`,
+});
 
 const generateArtistQuestionAnswers = (
   artists: string[]
@@ -11,4 +19,4 @@ const generateArtistQuestionAnswers = (
   return artistQuestionAnswers;
 };
 
-export { generateArtistQuestionAnswers };
+export { generateArtistQuestionAnswer, generateArtistQuestionAnswers };
