+++ src/mocks/helpers/generate-genre-question-answers.helper.ts
@@ -1,5 +1,9 @@
-import { generateGenreQuestionAnswer } from '~/mocks/helpers/generate-genre-question-answer.helper';
 import { IGenreQuestionAnswer } from '~/common/interfaces/interfaces';
+
+const generateGenreQuestionAnswer = (genre: string): IGenreQuestionAnswer => ({
+  genre,
+  src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
+});
 
 const generateGenreQuestionAnswers = (
   genres: string[]
@@ -13,4 +17,4 @@ const generateGenreQuestionAnswers = (
   return genreQuestionAnswers;
 };
 
-export { generateGenreQuestionAnswers };
+export { generateGenreQuestionAnswer, generateGenreQuestionAnswers };
