+++ src/mocks/helpers/helpers.ts
@@ -1,10 +1,6 @@
 export * from './generate-questions.helper';
-export * from './generate-question.helper';
 export * from './get-unique-answer-types.helper';
 export * from './generate-genre-question.helper';
-export * from './generate-genre-question-answer.helper';
 export * from './generate-genre-question-answers.helper';
 export * from './generate-artist-question.helper';
-export * from './generate-artist-question-song.helper';
-export * from './generate-artist-question-answer.helper';
 export * from './generate-artist-question-answers.helper';
